+++ src/i18n/index.ts
@@ -14,6 +14,8 @@ import ko from './ko'
 import zh from './zh'
 import fa from './fa'
 import sv from './sv'
+import cs from './cs'
+import it from './it'
 
 import ConsoleLogger from '~/logging/ConsoleLogger'
 import { TiptapVuetifyPlugin } from '~/main'
@@ -35,7 +37,9 @@ export const dictionary = {
   ko,
   zh,
   fa,
-  sv
+  sv,
+  cs,
+  it
 }
 
 export function getCurrentLang () {
@@ -46,9 +50,9 @@ export function getMsg (path: string, args?, lang: null | string = null): string
   let currentLang = lang || getCurrentLang()
 
   if (!dictionary[currentLang]) {
-    currentLang = defaultLanguage
-
     ConsoleLogger.warn(`The current language "${currentLang}" is not yet available. Using language "${defaultLanguage}" by default. Contribution to github is welcome.`)
+
+    currentLang = defaultLanguage
   }
 
   const dictionaryByLang = dictionary[currentLang]